+++ http/controller/api/peer.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"github.com/gin-gonic/gin"
 	"github.com/gin-gonic/gin/binding"
-	requstform "github.com/lejianwen/rustdesk-api/v2/http/request/api"
+	requestform "github.com/lejianwen/rustdesk-api/v2/http/request/api"
 	"github.com/lejianwen/rustdesk-api/v2/http/response"
 	"github.com/lejianwen/rustdesk-api/v2/service"
 	"net/http"
@@ -19,15 +19,15 @@ type Peer struct {
 // @Description 提交系统信息
 // @Accept  json
 // @Produce  json
-// @Param body body requstform.PeerForm true "系统信息表单"
+// @Param body body requestform.PeerForm true "系统信息表单"
 // @Success 200 {string} string "SYSINFO_UPDATED,ID_NOT_FOUND"
 // @Failure 500 {object} response.ErrorResponse
 // @Router /sysinfo [post]
 func (p *Peer) SysInfo(c *gin.Context) {
-	f := &requstform.PeerForm{}
+	f := &requestform.PeerForm{}
 	err := c.ShouldBindBodyWith(f, binding.JSON)
 	if err != nil {
-		response.Error(c, response.TranslateMsg(c, "ParamsError")+err.Error())
+		response.ErrorErr(c, "ParamsError", err)
 		return
 	}
 	fpe := f.ToPeer()
@@ -37,7 +37,7 @@ func (p *Peer) SysInfo(c *gin.Context) {
 		pe.UserId = service.AllService.UserService.FindLatestUserIdFromLoginLogByUuid(pe.Uuid, pe.Id)
 		err = service.AllService.PeerService.Create(pe)
 		if err != nil {
-			response.Error(c, response.TranslateMsg(c, "OperationFailed")+err.Error())
+			response.ErrorErr(c, "OperationFailed", err)
 			return
 		}
 	} else {
@@ -48,7 +48,7 @@ func (p *Peer) SysInfo(c *gin.Context) {
 		fpe.UserId = pe.UserId
 		err = service.AllService.PeerService.Update(fpe)
 		if err != nil {
-			response.Error(c, response.TranslateMsg(c, "OperationFailed")+err.Error())
+			response.ErrorErr(c, "OperationFailed", err)
 			return
 		}
 	}